Fix warning from incorrect qHash signature
The qHash prototype returns a size_t, not an int.

Pick-to: 6.1

+++ tests/auto/core5/tools/qlinkedlist/tst_qlinkedlist.cpp
@@ -86,7 +86,7 @@ Q_DECLARE_METATYPE(Movable);
 
 Q_DECLARE_METATYPE(QLinkedList<int>);
 
-int qHash(const Movable &movable)
+size_t qHash(const Movable &movable)
 {
     return qHash(movable.i);
 }
